Add early uart from rpi5 and use it as a console io
Also fix bug when there was no output in the console. The reason is that core::fmt::Write has its
own write_char imlementation which was used instead of PL011EarlyUartInner's one. In order to fix
it a Uart trait was introduced and Uart::write_char syntax used to guarantee that proper write_char
imlementation will be used