make uart on 14/15 pin work (also early uart)
Made uart on 14/15 work but the problem is that this is also "already configured" uart, if enable_rp1_uart=1 and pciex4_reset=0 present in the config.txt.
That means that this uart already configured by firmware and we do not need to setup gpio driver.
Also special address 0x1c00030000 must be used in order to make it work.

In order to make 14/15 uart work via fully configured driver, pcie bus must be configured (yet another driver) - this will be done later (pi5_hack repo will be used as an example)"